requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
allure-pytest==2.9.45
allure-python-commons==2.9.45; python_version >= '3.5'
async-generator==1.10; python_version >= '3.5'
atomicwrites==1.4.0; sys_platform == 'win32'
attrs==21.4.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
certifi==2021.10.8
cffi==1.15.0; os_name == 'nt' and implementation_name != 'pypy'
charset-normalizer==2.0.12; python_version >= '3'
colorama==0.4.4; sys_platform == 'win32'
cryptography==36.0.2
execnet==1.9.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
faker==13.3.4
furl==2.1.3
h11==0.13.0; python_version >= '3.6'
idna==3.3; python_version >= '3'
iniconfig==1.1.1
orderedmultidict==1.0.1
outcome==1.1.0; python_version >= '3.6'
packaging==21.3; python_version >= '3.6'
pluggy==1.0.0; python_version >= '3.6'
py==1.11.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
pycparser==2.21
pyopenssl==22.0.0
pyparsing==3.0.7; python_version >= '3.6'
pysocks==1.7.1
pytest-faker==2.0.0
pytest-forked==1.4.0; python_version >= '3.6'
pytest-repeat==0.9.1
pytest-xdist==2.5.0
pytest==7.1.1
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==0.20.0
requests==2.27.1
selenium==4.1.3
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.2.0; python_version >= '3.5'
sortedcontainers==2.4.0
tomli==2.0.1; python_version >= '3.7'
trio-websocket==0.9.2; python_version >= '3.5'
trio==0.20.0; python_version >= '3.7'
urllib3[secure,socks]==1.26.9; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
webdriver-manager==3.5.4
wsproto==1.1.0; python_version >= '3.7'
